// frontend/ReceiptTracker/src/app/api/model/userOut.ts
/**
 * Receipt Tracker API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserOut { 
    id: number;
    username: string;
    email?: string | null;
    fullname?: string | null;
    profile_picture?: string | null;
    disabled?: boolean | null;
    roles?: Array<string>;
}