// frontend/ReceiptTracker/src/app/api/model/wordCloudItem.ts
/**
 * Receipt Tracker API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WordCloudItem { 
    text: string;
    value: number;
    total_spent: number;
}