// frontend/ReceiptTracker/src/app/api/model/timeSeriesData.ts
/**
 * Receipt Tracker API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ModelDate } from './modelDate';


export interface TimeSeriesData { 
    date: ModelDate;
    value: number;
}